refactor(app/inbound): request body metrics submodule
